requirements.txt
# Telegram Channel Monitor - Requirements
# ========================================

# Core - Telegram
pyrogram>=2.0.106
tgcrypto>=1.2.5

# Environment
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Web Dashboard
flask>=3.0.0
flask-login>=0.6.3
flask-wtf>=1.2.1

# API
flask-restful>=0.3.10

# Scheduling
apscheduler>=3.10.4

# Notifications
aiohttp>=3.9.0
requests>=2.31.0

# Data Export
pandas>=2.1.0

# Logging
colorlog>=6.8.0

# Async Support
nest-asyncio>=1.5.8

# Email
aiosmtplib>=3.0.1

# Utils
python-dateutil>=2.8.2
